refactor(ui): move login logic to redux asyncThunk
