+++ packages/ui/src/pages/auth/ui/LoginPage.tsx
@@ -1,7 +1,8 @@
 import { Button, Form, Input, message, Typography } from 'antd';
 import { useState } from 'react';
-import { useLogin, useLogout } from 'src/features/auth';
+import { login, useLogout } from 'src/features/auth';
 import { AuthenticationError } from 'src/shared/lib';
+import { useAppDispatch } from 'src/shared/model';
 
 import { FormData } from '../model/FormData';
 import AuthForm from './AuthForm';
@@ -10,8 +11,8 @@ import AuthPageLayout from './AuthPageLayout';
 const { Text, Link } = Typography;
 
 export default function LoginPage() {
-  const login = useLogin();
   const logout = useLogout();
+  const dispatch = useAppDispatch();
   const [isLoading, setIsLoading] = useState(false);
   const [messageApi, contextHolder] = message.useMessage();
 
@@ -24,7 +25,7 @@ export default function LoginPage() {
       formParams.append('username', data.email);
       formParams.append('password', data.password);
 
-      await login(formParams);
+      dispatch(login(formParams));
     } catch (error) {
       const msgFailedReason =
         error instanceof AuthenticationError
